requirements.txt
numpy
aiofiles
aiogram
pillow
torch
torchvision
tensorflow
tensorflow_hub
matplotlib
imageio
